move error to module root
