+++ Cargo.toml
@@ -46,7 +46,7 @@ derive_builder = "0.20.0"
 ecdsa = "0.16.9"
 futures = "0.3.30"
 insta = { version = "1.39.0", features = ["redactions", "yaml"] }
-serde = { version = "1.0.204", features = ["derive"] }
+serde = { version = "1.0.205", features = ["derive"] }
 serde_json = { version = "1.0.122", features = ["alloc"] }
 serde_qs = "0.13.0"
 tokio = { version = "1.39.2", features = ["macros", "rt-multi-thread"] }
